Created new shared_communication_channel and moved shared global state into it. This will make it explicity that the communication betweeen Numbers and AutomaticDifferentiator is shared data over a communication channel and that this is by design
