+++ src/main.rs
@@ -4,6 +4,7 @@ pub mod automatic_differentiator;
 pub mod global_counter;
 pub mod number;
 pub mod operation;
+pub mod shared_data_communication_channel;
 
 fn g(args: Vec<Number>) -> Number {
     let x1 = args[0];
