+++ src/number.rs
@@ -9,6 +9,7 @@ use crate::automatic_differentiator;
 use crate::operation::Operation;
 
 use crate::global_counter::OPERATION_ID_COUNTER;
+use crate::shared_data_communication_channel;
 
 #[derive(Debug, Clone, Copy)]
 pub struct Number {
@@ -45,19 +46,19 @@ impl Add for Number {
     fn add(self, rhs: Self) -> Self::Output {
         let result: Number = Number::new_non_leaf(self.result + rhs.result);
         let op = Operation::Add(result.id, self.id, rhs.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
             result.id,
             vec![self.id, rhs.id],
         );
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         if rhs.leaf {
             let val_op = Operation::Value(rhs.id, rhs.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
         result
     }
@@ -69,20 +70,20 @@ impl Sub for Number {
     fn sub(self, rhs: Self) -> Self::Output {
         let result: Number = Number::new_non_leaf(self.result - rhs.result);
         let op = Operation::Sub(result.id, self.id, rhs.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
             result.id,
             vec![self.id, rhs.id],
         );
 
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         if rhs.leaf {
             let val_op = Operation::Value(rhs.id, rhs.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
         result
     }
@@ -94,19 +95,19 @@ impl Mul for Number {
     fn mul(self, rhs: Self) -> Self::Output {
         let result: Number = Number::new_non_leaf(self.result * rhs.result);
         let op = Operation::Mul(result.id, self.id, rhs.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
             result.id,
             vec![self.id, rhs.id],
         );
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         if rhs.leaf {
             let val_op = Operation::Value(rhs.id, rhs.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         result
@@ -119,20 +120,20 @@ impl Div for Number {
     fn div(self, rhs: Self) -> Self::Output {
         let result: Number = Number::new_non_leaf(self.result / rhs.result);
         let op = Operation::Div(result.id, self.id, rhs.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
             result.id,
             vec![self.id, rhs.id],
         );
 
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         if rhs.leaf {
             let val_op = Operation::Value(rhs.id, rhs.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         result
@@ -143,12 +144,15 @@ impl Number {
     pub fn ln(self) -> Number {
         let result: Number = Number::new_non_leaf(self.result.ln());
         let op = Operation::Ln(result.id, self.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(result.id, vec![self.id]);
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
+            result.id,
+            vec![self.id],
+        );
 
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         result
@@ -157,12 +161,15 @@ impl Number {
     pub fn sin(self) -> Number {
         let result: Number = Number::new_non_leaf(self.result.sin());
         let op = Operation::Sin(result.id, self.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(result.id, vec![self.id]);
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
+            result.id,
+            vec![self.id],
+        );
 
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         result
@@ -171,12 +178,15 @@ impl Number {
     pub fn exp(self) -> Number {
         let result: Number = Number::new_non_leaf(self.result.exp());
         let op = Operation::Exp(result.id, self.id, result.result, 0.0);
-        automatic_differentiator::global_register_operation(op);
-        automatic_differentiator::global_add_parent_child_relationship(result.id, vec![self.id]);
+        shared_data_communication_channel::global_register_operation(op);
+        shared_data_communication_channel::global_add_parent_child_relationship(
+            result.id,
+            vec![self.id],
+        );
 
         if self.leaf {
             let val_op = Operation::Value(self.id, self.result, 0.0);
-            automatic_differentiator::global_register_operation(val_op);
+            shared_data_communication_channel::global_register_operation(val_op);
         }
 
         result
